Add Delete Function to Listings
Now, users can delete their listings by entering their email in a form. This feature is built on the same lines of the Edit Listing and has the same flow of actions. First the user requests deletion and then he/she confirms it by following the link sent in the email.